+++ libraries/stdlib/native-wasm/src/kotlin/text/unicode/CodePointCasing.kt
@@ -3,23 +3,21 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package kotlin.text.codepoints
+package kotlin.text.unicode
+
 
 @ExperimentalCodePointApi
 public actual fun CodePoint.lowercase(): String =
-    toString().lowercase()
+    if (isBasic) code.toChar().lowercaseImpl() else lowercaseCodePoint().toString()
 
 @ExperimentalCodePointApi
 public actual fun CodePoint.lowercaseCodePoint(): CodePoint =
-    // there's a single multichar expansion
-    lowercase().codePointAt(0)
+    lowercaseCodePoint(code).toCodePoint()
 
 @ExperimentalCodePointApi
 public actual fun CodePoint.uppercase(): String =
-    toString().uppercase()
+    if (isBasic) code.toChar().uppercaseImpl() else uppercaseCodePoint().toString()
 
 @ExperimentalCodePointApi
-public actual fun CodePoint.uppercaseCodePoint(): CodePoint {
-    val uppercase = uppercase()
-    return if (uppercase.codePointCount() > 1) this else uppercase.codePointAt(0)
-}
+public actual fun CodePoint.uppercaseCodePoint(): CodePoint =
+    uppercaseCodePoint(code).toCodePoint()
