+++ libraries/stdlib/src/kotlin/text/unicode/CodePointCasing.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package kotlin.text.codepoints
+package kotlin.text.unicode
 
 /**
  * Converts this code point to lower case using Unicode mapping rules of the invariant locale.
