+++ libraries/stdlib/test/text/CodePointRangeTest.kt
@@ -5,7 +5,7 @@
 
 package test.text
 
-import kotlin.text.codepoints.*
+import kotlin.text.unicode.*
 import kotlin.test.*
 
 @OptIn(ExperimentalCodePointApi::class)
